rustup file must exist

+++ install.sh
@@ -288,6 +288,7 @@ warnSudo(){
 
 cargoinstall(){
     prettyPrint "Installing rustup for exa, fd and bat in background"
+    test -f rustupinstall.sh || { echo "Where is rustupinstall.sh in zpwr base directory?" >&2; exit 1; }
     bash rustupinstall.sh &> "$logfile_cargo" &
     CARGO_PID=$!
     trap 'kill $CARGO_PID 2>/dev/null;echo bye;exit' INT
